Crate lapack_src

LAPACK source of choice.

The usage of the package is explained here.

Configuration

The following implementations are available:

accelerate, which is the one in the Accelerate framework (macOS only),

intel-mkl, which is the one in Intel MKL,

netlib, which is the reference one by Netlib, and

openblas, which is the one in OpenBLAS.

An implementation can be chosen as follows:

[dependencies]
lapack-src = { version = "0.4", features = ["accelerate"] }
lapack-src = { version = "0.4", features = ["intel-mkl"] }
lapack-src = { version = "0.4", features = ["netlib"] }
lapack-src = { version = "0.4", features = ["openblas"] }